updates for 64-bit support

+++ ulam1/include/Parser.h
@@ -393,8 +393,8 @@ namespace MFM{
        helper method to make a terminal node
        for a constant value known at parse time (e.g. one for ++/--)
     */
-    Node * makeTerminal(Token& locTok, s32 val, UTI utype);
-    Node * makeTerminal(Token& locTok, u32 val, UTI utype);
+    Node * makeTerminal(Token& locTok, s64 val, UTI utype);
+    Node * makeTerminal(Token& locTok, u64 val, UTI utype);
 
     /**
        helper method to save subtrees for unknown UTIs
